Adding headless tests for CASUALcore Reworking parts of CASUAL to run in headless mode Renaming PACKAGER to CASPACkager Upgrading to netbeans 7.4 for CASCADE, CASPACcreator, and CASPACkager Fixing remaining "compiling 1.6 source base under 1.7 compiler" errors. Adding catch-all create new file for zip.addToExistingZip Fixing CASPACcreator tests Fixing CASPACkager tests

+++ CASUALcore/test/CASUAL/misc/JarResources.java
@@ -80,11 +80,11 @@ public final class JarResources {
                     size = ((Integer) htSizes.get(ze.getName())).intValue();
                 }
 
-                byte[] b = new byte[(int) size];
+                byte[] b = new byte[size];
                 int rb = 0;
                 int chunk;
-                while (((int) size - rb) > 0) {
-                    chunk = zis.read(b, rb, (int) size - rb);
+                while (( size - rb) > 0) {
+                    chunk = zis.read(b, rb, size - rb);
                     if (chunk == -1) {
                         break;
                     }
